VEOSS-692 | Moved SessionRepository clean from CheckYourAnswersController to ReturnSubmittedController
