+++ test/controllers/CheckYourAnswersControllerSpec.scala
@@ -26,7 +26,7 @@ import models.domain.VatReturn
 import models.emails.EmailSendingResult.EMAIL_ACCEPTED
 import models.requests.DataRequest
 import models.responses.{ConflictFound, UnexpectedResponseStatus}
-import models.{Country, NormalMode, TotalVatToCountry, UserAnswers}
+import models.{Country, NormalMode, TotalVatToCountry}
 import org.mockito.ArgumentMatchers.any
 import org.mockito.ArgumentMatchersSugar.eqTo
 import org.mockito.Mockito
@@ -37,7 +37,6 @@ import org.scalatestplus.mockito.MockitoSugar
 import pages.corrections.{CorrectPreviousReturnPage, CorrectionCountryPage, CorrectionReturnPeriodPage, CountryVatCorrectionPage}
 import pages.{CheckYourAnswersPage, SoldGoodsFromEuPage, SoldGoodsFromNiPage}
 import play.api.inject.bind
-import play.api.libs.json.Json
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
 import queries.EmailConfirmationQuery
@@ -46,7 +45,6 @@ import services.corrections.CorrectionService
 import services.{AuditService, EmailService, SalesAtVatRateService, VatReturnService}
 import viewmodels.govuk.SummaryListFluency
 
-import java.time.{Clock, Instant, ZoneId}
 import scala.concurrent.Future
 
 class CheckYourAnswersControllerSpec extends SpecBase with MockitoSugar with SummaryListFluency with BeforeAndAfterEach {
@@ -229,7 +227,6 @@ class CheckYourAnswersControllerSpec extends SpecBase with MockitoSugar with Sum
 
           val answers = completeUserAnswers
 
-          when(mockSessionRepository.clear(any())) thenReturn Future.successful(true)
           when(mockSessionRepository.set(any())) thenReturn Future.successful(true)
           when(vatReturnService.fromUserAnswers(any(), any(), any(), any())) thenReturn Valid(vatReturnRequest)
           when(vatReturnConnector.submit(any())(any())) thenReturn Future.successful(Right(vatReturn))
@@ -261,8 +258,7 @@ class CheckYourAnswersControllerSpec extends SpecBase with MockitoSugar with Sum
             )
             val expectedAuditEventForDataEntry = ReturnForDataEntryAuditModel(vatReturnRequest, None, vatReturn.reference, vatReturn.paymentReference)
 
-            val userAnswersWithEmailConfirmation = answers.copy(data = Json.obj())
-              .set(EmailConfirmationQuery, true).success.value
+            val userAnswersWithEmailConfirmation = completeUserAnswers.copy().set(EmailConfirmationQuery, true).success.value
 
             status(result) mustEqual SEE_OTHER
             redirectLocation(result).value mustEqual CheckYourAnswersPage.navigate(NormalMode, userAnswersWithEmailConfirmation).url
@@ -276,7 +272,6 @@ class CheckYourAnswersControllerSpec extends SpecBase with MockitoSugar with Sum
                 eqTo(totalVatOnSales),
                 eqTo(vatReturnRequest.period)
               )(any(), any())
-            verify(mockSessionRepository, times(1)).clear(eqTo(userAnswersWithEmailConfirmation.userId))
             verify(mockSessionRepository, times(1)).set(eqTo(userAnswersWithEmailConfirmation))
           }
         }
@@ -325,7 +320,6 @@ class CheckYourAnswersControllerSpec extends SpecBase with MockitoSugar with Sum
           when(vatReturnConnector.submitWithCorrection(any())(any())) thenReturn Future.successful(Right(vatReturn, correctionPayload))
           when(emailService.sendConfirmationEmail(any(), any(), any(), any(), any())(any(), any()))
             .thenReturn(Future.successful(EMAIL_ACCEPTED))
-          when(mockSessionRepository.clear(any())) thenReturn Future.successful(true)
 
           val totalVatOnSales = BigDecimal(100)
           when(salesAtVatRateService.getTotalVatOwedAfterCorrections(any())) thenReturn totalVatOnSales
@@ -353,7 +347,7 @@ class CheckYourAnswersControllerSpec extends SpecBase with MockitoSugar with Sum
             )
             val expectedAuditEventForDataEntry = ReturnForDataEntryAuditModel(vatReturnRequest, Some(correctionRequest), vatReturn.reference, vatReturn.paymentReference)
 
-            val userAnswersWithEmailConfirmation = answers.copy(data = Json.obj()).set(EmailConfirmationQuery, true).success.value
+            val userAnswersWithEmailConfirmation = answers.set(EmailConfirmationQuery, true).success.value
 
             status(result) mustEqual SEE_OTHER
             redirectLocation(result).value mustEqual CheckYourAnswersPage.navigate(NormalMode, userAnswersWithEmailConfirmation).url
@@ -493,38 +487,6 @@ class CheckYourAnswersControllerSpec extends SpecBase with MockitoSugar with Sum
         contentAsString(result).contains("&pound;3,333") mustBe true
         contentAsString(result).contains("&pound;7,777") mustBe true
         contentAsString(result).contains("&pound;8,888") mustBe true
-      }
-    }
-
-    "must clear user-answers when return is submitted" in {
-      val mockSessionRepository = mock[SessionRepository]
-
-      val answers = completeUserAnswers
-
-      when(vatReturnConnector.submit(any())(any())) thenReturn Future.successful(Right(vatReturn))
-      when(emailService.sendConfirmationEmail(any(), any(), any(), any(), any())(any(), any()))
-        .thenReturn(Future.successful(EMAIL_ACCEPTED))
-
-      when(mockSessionRepository.clear(any())) thenReturn Future.successful(true)
-      when(mockSessionRepository.set(any())) thenReturn Future.successful(true)
-
-      val app = applicationBuilder(userAnswers = Some(answers))
-        .overrides(
-          bind[SessionRepository].toInstance(mockSessionRepository),
-          bind[VatReturnConnector].toInstance(vatReturnConnector),
-          bind[CorrectionConnector].toInstance(correctionConnector),
-          bind[EmailService].toInstance(emailService)
-        ).build()
-
-      running(app) {
-
-        val request = FakeRequest(POST, routes.CheckYourAnswersController.onSubmit(period).url)
-
-        val result = route(app, request).value
-
-        status(result) mustEqual SEE_OTHER
-        redirectLocation(result).value mustEqual routes.ReturnSubmittedController.onPageLoad(period).url
-        verify(mockSessionRepository, times(1)).clear(eqTo(answers.userId))
       }
     }
   }
